+++ strelyaev.roman/P4/main.cpp
@@ -19,7 +19,28 @@ void checkArgs(int argc, int n, std::fstream & in)
     throw (std::logic_error("Unable to read input file"));
   }
 }
+void sMatrix (std::istream & in, int * a, int e)
+{
+  for (int i = 0; i < e; i++)
+  {
+    if(!(in >> a[i]))
+    {
+      throw (std::logic_error("Unable to create matrix"));
+    }
+  }
+}
 
+void dMatrix (std::istream & in, int * a, int e)
+{
+  for (int i = 0; i < e; i++)
+  {
+    if(!(in >> a[i]))
+    {
+      delete [] a;
+      throw (std::logic_error("Unable to create matrix"));
+    }
+  }
+}
 void matrixSubtraction(int * original, int * reverse, int n)
 {
   for (int i = 0; i < n; i++)
