--- strelyaev.roman/P4/matrix.h
@@ -1,6 +1,0 @@
-#ifndef MATRIX_H
-#define MATRIX_H
-#include <fstream>
-void sMatrix (std::istream & in, int * a, int e);
-void dMatrix (std::istream & in, int * a, int e);
-#endif
